Add C11 preprocessor arch conditionals, rewrite string.plib
- Add #ifdef __ARM64__ / __x86_64__ preprocessor support
- Rewrite plstd/string.plib with architecture conditionals
- Remove asm_common.py, move code into assembler.py and nasmgen.py
- Add preprocessor evaluation in main.py (preprocess_source)
- Fix import issues in assembler.py
- String functions now conditionally compile based on target arch
