+++ src/nasmgen.py
@@ -15,11 +15,471 @@ import platform
 import subprocess
 from typing import List, Optional, Tuple
 
-import asm_common
+X86_INT_REGS = ["rdi", "rsi", "rdx", "rcx", "r8", "r9"]
+X86_FLOAT_REGS = ["xmm0", "xmm1", "xmm2", "xmm3", "xmm4", "xmm5", "xmm6", "xmm7"]
+SYS_WRITE_X86 = 1
+SYS_READ_X86 = 0
+SYS_OPEN_X86 = 2
+SYS_CLOSE_X86 = 3
+SYS_EXIT_X86 = 60
+SYS_WRITE_ARM64_MACOS = 4
+SYS_READ_ARM64_MACOS = 3
+SYS_OPEN_ARM64_MACOS = 5
+SYS_CLOSE_ARM64_MACOS = 6
+SYS_EXIT_ARM64_MACOS = 1
+
+def emit_syscall(
+    syscall_num: int,
+    args: Optional[List[str]] = None,
+    is_arm64: bool = False,
+    is_macos: bool = False,
+):
+    """Emit syscall instruction sequence for current architecture."""
+    if args is None:
+        args = []
+
+    if is_arm64:
+        if is_macos:
+            lines = [f"mov x16, #{syscall_num}"]
+            for i, arg in enumerate(args[:6]):
+                lines.append(f"mov x{i}, {arg}")
+            lines.append("svc #0x80")
+            return lines
+        else:
+            lines = [f"mov x8, {syscall_num}"]
+            for i, arg in enumerate(args[:6]):
+                lines.append(f"mov x{i}, {arg}")
+            lines.append("svc #0")
+            return lines
+    else:
+        lines = [f"mov rax, {syscall_num}"]
+        for i, arg in enumerate(args[:6]):
+            reg = X86_INT_REGS[i]
+            lines.append(f"mov {reg}, {arg}")
+        lines.append("syscall")
+        return lines
+
+
+def get_asm_config(
+    asm_block_syntax: Optional[str],
+    target_arch: Optional[str] = None,
+    output_format: Optional[str] = None,
+    is_macos: bool = False,
+    current_arch: Optional[str] = None,
+):
+    """Determine architecture and format for an asm block."""
+    is_arm64_current = False
+    format_current = "elf64"
+
+    if target_arch == "arm64":
+        is_arm64_current = True
+        format_current = "macho64" if is_macos else "elf64"
+    elif target_arch == "x86_64":
+        is_arm64_current = False
+        format_current = "macho64" if is_macos else "elf64"
+
+    if output_format == "macho":
+        is_arm64_current = True
+        format_current = "macho64"
+    elif output_format == "elf":
+        is_arm64_current = False
+        format_current = "elf64"
+
+    if asm_block_syntax:
+        syntax = asm_block_syntax.lower()
+        if "arm64" in syntax:
+            is_arm64_current = True
+            if "macho" in syntax:
+                format_current = "macho64"
+            else:
+                format_current = "elf64"
+        elif "x86_64" in syntax:
+            is_arm64_current = False
+            if "macho" in syntax:
+                format_current = "macho64"
+            else:
+                format_current = "elf64"
+
+    if not target_arch and not output_format and not asm_block_syntax:
+        if is_macos:
+            if current_arch == "arm64":
+                is_arm64_current = True
+                format_current = "macho64"
+            else:
+                is_arm64_current = False
+                format_current = "macho64"
+        else:
+            is_arm64_current = False
+            format_current = "elf64"
+
+    return is_arm64_current, format_current
+
+
+def _escape_string(value: str) -> str:
+    """Escape a string for assembly."""
+    return (
+        value.replace("\\", "\\\\")
+        .replace("\n", "\\n")
+        .replace("\t", "\\t")
+        .replace('"', '\\"')
+    )
+
+
+def _expand_write_call(instr: str, is_arm64: bool) -> List[str]:
+    """Replace _write calls with pure syscalls."""
+    stripped = instr.strip()
+    if "bl" in stripped.lower() and "_write" in stripped:
+        if is_arm64:
+            return ["mov x16, #4", "svc #0x80"]
+        else:
+            return ["mov rax, 1", "syscall"]
+    return [instr]
+
+
+def _expand_len_instruction(instr: str, is_arm64: bool, label_suffix: str = "") -> List[str]:
+    """Expand len() into inline assembly.
+
+    Args:
+        instr: The instruction string to process
+        is_arm64: True for ARM64, False for x86_64
+        label_suffix: Optional suffix for labels to make them unique
+    """
+    match = re.match(r"^\s*mov\s+([^,]+),\s*len\(([^)]+)\)\s*$", instr)
+    if not match:
+        return [instr]
+
+    dest = match.group(1).strip()
+    src = match.group(2).strip()
+
+    if label_suffix:
+        loop_label = f".Lhyp_len_loop_{label_suffix}" if is_arm64 else f".hyp_len_loop_{label_suffix}"
+        done_label = f".Lhyp_len_done_{label_suffix}" if is_arm64 else f".hyp_len_done_{label_suffix}"
+    else:
+        loop_label = ".Lhyp_len_loop" if is_arm64 else ".hyp_len_loop"
+        done_label = ".Lhyp_len_done" if is_arm64 else ".hyp_len_done"
+
+    if is_arm64:
+        return [
+            f"mov x10, {src}",
+            f"mov {dest}, #0",
+            f"{loop_label}:",
+            f"ldrb w11, [x10, {dest}]",
+            f"cbz w11, {done_label}",
+            f"add {dest}, {dest}, #1",
+            f"b {loop_label}",
+            f"{done_label}:",
+        ]
+    else:
+        return [
+            f"mov r10, {src}",
+            f"xor {dest}, {dest}",
+            f"{loop_label}:",
+            f"cmp byte [r10 + {dest}], 0",
+            f"je {done_label}",
+            f"inc {dest}",
+            f"jmp {loop_label}",
+            f"{done_label}:",
+        ]
+
+
+def _split_binary_expr(expr: str):
+    parts = expr.split()
+    if len(parts) == 3 and parts[1] in ("+", "-", "*", "/", "%"):
+        return parts[0], parts[1], parts[2]
+    return None
+
+
+def _emit_integer_expr_x86(f, expr: str, dst: str):
+    binary = _split_binary_expr(expr)
+    if not binary:
+        f.write(f"    mov {dst}, {expr}\n")
+        return
+    left, op, right = binary
+    f.write(f"    mov {dst}, {left}\n")
+    if op == "+":
+        f.write(f"    add {dst}, {right}\n")
+    elif op == "-":
+        f.write(f"    sub {dst}, {right}\n")
+    elif op == "*":
+        f.write(f"    imul {dst}, {right}\n")
+    elif op in ("/", "%"):
+        f.write("    cqo\n")
+        f.write(f"    mov r10, {right}\n")
+        f.write("    idiv r10\n")
+        if op == "%":
+            f.write(f"    mov {dst}, rdx\n")
+
+
+def _emit_integer_expr_arm64(f, expr: str, dst: str):
+    binary = _split_binary_expr(expr)
+    if not binary:
+        f.write(f"    mov {dst}, {expr}\n")
+        return
+    left, op, right = binary
+    f.write(f"    mov {dst}, {left}\n")
+    if op == "+":
+        f.write(f"    add {dst}, {dst}, {right}\n")
+    elif op == "-":
+        f.write(f"    sub {dst}, {dst}, {right}\n")
+    elif op == "*":
+        f.write(f"    mul {dst}, {dst}, {right}\n")
+    elif op in ("/", "%"):
+        f.write(f"    sdiv x9, {dst}, {right}\n")
+        if op == "/":
+            f.write(f"    mov {dst}, x9\n")
+        else:
+            f.write(f"    msub {dst}, x9, {right}, {dst}\n")
+
+
+def _emit_return_expr(f, return_expr, ret_type, is_arm64: bool):
+    """Emit instructions that place a simple return expression in the ABI register."""
+    if return_expr is None or return_expr == "":
+        return
+    if ret_type in ("float", "double"):
+        dst = "v0" if is_arm64 else "xmm0"
+        op = "fmov" if is_arm64 else "movq"
+        f.write(f"    {op} {dst}, {return_expr}\n")
+        return
+    if is_arm64:
+        _emit_integer_expr_arm64(f, return_expr, "x0")
+    else:
+        _emit_integer_expr_x86(f, return_expr, "rax")
+
+
+def _build_param_map(
+    params: List[Tuple[str, str]],
+    int_regs: List[str],
+    float_regs: List[str],
+    is_arm64: bool,
+):
+    """Build parameter to register mapping based on ABI."""
+    param_map = {}
+    int_idx = 0
+    float_idx = 0
+    offset = 8 if not is_arm64 else 0
+
+    for param_type, param_name in params:
+        if param_type in ("float", "double"):
+            if float_idx < len(float_regs):
+                param_map[param_name] = float_regs[float_idx]
+                float_idx += 1
+            else:
+                if is_arm64:
+                    param_map[param_name] = f"[sp, #{offset}]"
+                else:
+                    param_map[param_name] = f"[rbp-{offset}]"
+                offset += 8
+        else:
+            if int_idx < len(int_regs):
+                param_map[param_name] = int_regs[int_idx]
+                int_idx += 1
+            else:
+                if is_arm64:
+                    param_map[param_name] = f"[sp, #{offset}]"
+                else:
+                    param_map[param_name] = f"[rbp-{offset}]"
+                offset += 8
+
+    return param_map
+
+
+def _process_instructions(
+    instructions: List[str], param_map: dict, is_arm64: bool,
+):
+    """Process instructions, replacing parameter names with registers."""
+    updated = []
+    for instr in instructions:
+        result = instr
+        for param_name, addr in param_map.items():
+            pattern = r"\b" + re.escape(param_name) + r"\b"
+            result = re.sub(pattern, addr, result)
+        updated.append(result)
+    return updated
+
 
 PLATFORM = platform.system()
 IS_MACOS = PLATFORM == "Darwin"
 IS_LINUX = PLATFORM == "Linux"
+
+X86_INT_REGS = ["rdi", "rsi", "rdx", "rcx", "r8", "r9"]
+X86_FLOAT_REGS = ["xmm0", "xmm1", "xmm2", "xmm3", "xmm4", "xmm5", "xmm6", "xmm7"]
+
+
+def get_asm_config(syntax: Optional[str]) -> Tuple[bool, str]:
+    """Determine architecture and format for an asm block.
+
+    NASM only supports x86_64, so is_arm64 is always False.
+    """
+    is_arm64 = False
+    if IS_MACOS:
+        format_str = "macho64"
+    else:
+        format_str = "elf64"
+
+    if syntax:
+        syntax_lower = syntax.lower()
+        if "macho" in syntax_lower:
+            format_str = "macho64"
+        elif "elf" in syntax_lower:
+            format_str = "elf64"
+
+    return is_arm64, format_str
+
+
+def _escape_string(value: str) -> str:
+    """Escape a string for assembly."""
+    return (
+        value.replace("\\", "\\\\")
+        .replace("\n", "\\n")
+        .replace("\t", "\\t")
+        .replace('"', '\\"')
+    )
+
+
+def _expand_write_call(instr: str, is_arm64: bool) -> List[str]:
+    """Replace _write calls with pure syscalls."""
+    stripped = instr.strip()
+    if "bl" in stripped.lower() and "_write" in stripped:
+        if is_arm64:
+            return ["mov x16, #4", "svc #0x80"]
+        else:
+            return ["mov rax, 1", "syscall"]
+    return [instr]
+
+
+def _expand_len_instruction(instr: str, is_arm64: bool, label_suffix: str = "") -> List[str]:
+    """Expand len() into inline assembly.
+
+    Args:
+        instr: The instruction string to process
+        is_arm64: True for ARM64, False for x86_64
+        label_suffix: Optional suffix for labels to make them unique
+    """
+    match = re.match(r"^\s*mov\s+([^,]+),\s*len\(([^)]+)\)\s*$", instr)
+    if not match:
+        return [instr]
+
+    dest = match.group(1).strip()
+    src = match.group(2).strip()
+
+    if label_suffix:
+        loop_label = f".Lhyp_len_loop_{label_suffix}" if is_arm64 else f".hyp_len_loop_{label_suffix}"
+        done_label = f".Lhyp_len_done_{label_suffix}" if is_arm64 else f".hyp_len_done_{label_suffix}"
+    else:
+        loop_label = ".Lhyp_len_loop" if is_arm64 else ".hyp_len_loop"
+        done_label = ".Lhyp_len_done" if is_arm64 else ".hyp_len_done"
+
+    if is_arm64:
+        return [
+            f"mov x10, {src}",
+            f"mov {dest}, #0",
+            f"{loop_label}:",
+            f"ldrb w11, [x10, {dest}]",
+            f"cbz w11, {done_label}",
+            f"add {dest}, {dest}, #1",
+            f"b {loop_label}",
+            f"{done_label}:",
+        ]
+    else:
+        return [
+            f"mov r10, {src}",
+            f"xor {dest}, {dest}",
+            f"{loop_label}:",
+            f"cmp byte [r10 + {dest}], 0",
+            f"je {done_label}",
+            f"inc {dest}",
+            f"jmp {loop_label}",
+            f"{done_label}:",
+        ]
+
+
+def _split_binary_expr(expr: str):
+    parts = expr.split()
+    if len(parts) == 3 and parts[1] in ("+", "-", "*", "/", "%"):
+        return parts[0], parts[1], parts[2]
+    return None
+
+
+def _emit_integer_expr_x86(f, expr: str, dst: str):
+    binary = _split_binary_expr(expr)
+    if not binary:
+        f.write(f"    mov {dst}, {expr}\n")
+        return
+    left, op, right = binary
+    f.write(f"    mov {dst}, {left}\n")
+    if op == "+":
+        f.write(f"    add {dst}, {right}\n")
+    elif op == "-":
+        f.write(f"    sub {dst}, {right}\n")
+    elif op == "*":
+        f.write(f"    imul {dst}, {right}\n")
+    elif op in ("/", "%"):
+        f.write("    cqo\n")
+        f.write(f"    mov r10, {right}\n")
+        f.write("    idiv r10\n")
+        if op == "%":
+            f.write(f"    mov {dst}, rdx\n")
+
+
+def _emit_return_expr(f, return_expr, ret_type):
+    """Emit instructions that place a simple return expression in the ABI register."""
+    if return_expr is None or return_expr == "":
+        return
+    if ret_type in ("float", "double"):
+        f.write(f"    movq xmm0, {return_expr}\n")
+        return
+    _emit_integer_expr_x86(f, return_expr, "rax")
+
+
+def _build_param_map(
+    params: List[Tuple[str, str]],
+    int_regs: List[str],
+    float_regs: List[str],
+    is_arm64: bool,
+) -> dict:
+    """Build parameter to register mapping based on ABI."""
+    param_map = {}
+    int_idx = 0
+    float_idx = 0
+    offset = 8 if not is_arm64 else 0
+
+    for param_type, param_name in params:
+        if param_type in ("float", "double"):
+            if float_idx < len(float_regs):
+                param_map[param_name] = float_regs[float_idx]
+                float_idx += 1
+            else:
+                if is_arm64:
+                    param_map[param_name] = f"[sp, #{offset}]"
+                else:
+                    param_map[param_name] = f"[rbp-{offset}]"
+                offset += 8
+        else:
+            if int_idx < len(int_regs):
+                param_map[param_name] = int_regs[int_idx]
+                int_idx += 1
+            else:
+                if is_arm64:
+                    param_map[param_name] = f"[sp, #{offset}]"
+                else:
+                    param_map[param_name] = f"[rbp-{offset}]"
+                offset += 8
+
+    return param_map
+
+
+def _process_instructions(
+    instructions: List[str], param_map: dict, is_arm64: bool
+) -> List[str]:
+    """Process instructions, replacing parameter names with registers."""
+    updated = []
+    for instr in instructions:
+        result = instr
+        for param_name, addr in param_map.items():
+            pattern = r"\b" + re.escape(param_name) + r"\b"
+            result = re.sub(pattern, addr, result)
+        updated.append(result)
+    return updated
 
 
 def generate_nasm_file(
@@ -64,15 +524,15 @@ def _generate_x86_asm(asm_block, f, symbol_prefix: str, is_macos: bool = False):
         else:
             instructions.append(stripped)
 
-    param_map = asm_common._build_param_map(
-        asm_block.params, asm_common.X86_INT_REGS, asm_common.X86_FLOAT_REGS, is_arm64=False
+    param_map = _build_param_map(
+        asm_block.params, X86_INT_REGS, X86_FLOAT_REGS, is_arm64=False
     )
 
-    updated_instructions = asm_common._process_instructions(
+    updated_instructions = _process_instructions(
         instructions, param_map, is_arm64=False
     )
 
-    data_lines = _build_data_section(asm_block.variables, symbol_prefix, is_arm64=False)
+    data_lines = _build_data_section(asm_block.variables, symbol_prefix)
     if data_lines:
         f.write("section .data\n")
         for dl in data_lines:
@@ -108,7 +568,7 @@ def _generate_x86_asm(asm_block, f, symbol_prefix: str, is_macos: bool = False):
                 continue
             f.write(f"    {instr}\n")
 
-        asm_common._emit_return_expr(f, asm_block.return_expr, asm_block.ret_type, is_arm64=False)
+        _emit_return_expr(f, asm_block.return_expr, asm_block.ret_type, is_arm64=False)
 
         f.write("    pop rbp\n")
         f.write("    ret\n")
@@ -116,12 +576,12 @@ def _generate_x86_asm(asm_block, f, symbol_prefix: str, is_macos: bool = False):
         for instr in updated_instructions:
             f.write(f"{instr}\n")
         if asm_block.return_expr is not None:
-            asm_common._emit_return_expr(f, asm_block.return_expr, asm_block.ret_type, is_arm64=False)
+            _emit_return_expr(f, asm_block.return_expr, asm_block.ret_type, is_arm64=False)
             f.write("    ret\n")
 
 
 def _build_data_section(
-    variables: List[dict], symbol_prefix: str, is_arm64: bool
+    variables: List[dict], symbol_prefix: str
 ) -> List[str]:
     """Build data section lines for variables."""
     lines = []
@@ -138,38 +598,22 @@ def _build_data_section(
         lines.append(f"global {symbol_name}")
         lines.append(f"{symbol_name}:")
 
-        if is_arm64:
-            if var_type == "string":
-                escaped = asm_common._escape_string(initializer or "")
-                lines.append(f'    .asciz "{escaped}"')
-            elif var_type == "char":
-                if array_size:
-                    lines.append(f"    .fill {array_size}, 1, {initializer or 0}")
-                else:
-                    lines.append(f"    .byte {initializer or 0}")
-            elif var_type in ("int", "long"):
-                lines.append(f"    .long {initializer or 0}")
-            elif var_type == "float":
-                lines.append(f"    .float {initializer or 0}")
-            elif var_type == "double":
-                lines.append(f"    .double {initializer or 0}")
+        type_to_nasm = {
+            "string": "db",
+            "char": "db",
+            "short": "dw",
+            "int": "dd",
+            "long": "dd",
+            "float": "dd",
+            "double": "dq",
+        }
+        nasm_directive = type_to_nasm.get(var_type, "dd")
+        if array_size:
+            lines.append(
+                f"    times {array_size} {nasm_directive} {initializer or 0}"
+            )
         else:
-            type_to_nasm = {
-                "string": "db",
-                "char": "db",
-                "short": "dw",
-                "int": "dd",
-                "long": "dd",
-                "float": "dd",
-                "double": "dq",
-            }
-            nasm_directive = type_to_nasm.get(var_type, "dd")
-            if array_size:
-                lines.append(
-                    f"    times {array_size} {nasm_directive} {initializer or 0}"
-                )
-            else:
-                lines.append(f"    {nasm_directive} {initializer or 0}")
+            lines.append(f"    {nasm_directive} {initializer or 0}")
 
         lines.append("")
 
@@ -187,7 +631,7 @@ def emit_nasm_file(asm_block, output_path: str, syntax: str) -> str:
     Returns:
         The output_path that was written to
     """
-    is_arm64, format_str = asm_common.get_asm_config(syntax)
+    is_arm64, format_str = get_asm_config(syntax)
     generate_nasm_file(asm_block, output_path, is_arm64, format_str)
     return output_path
 
@@ -216,7 +660,7 @@ def assemble_with_nasm(asm_path: str, obj_path: str, format_str: str) -> str:
         obj_path = base + ".o"
 
     if format_str is None:
-        is_arm64, format_str = asm_common.get_asm_config(None)
+        is_arm64, format_str = get_asm_config(None)
 
     result = subprocess.run(
         ["nasm", "-f", format_str, asm_path, "-o", obj_path],
@@ -241,7 +685,7 @@ def emit_and_assemble(asm_block, output_path: str, syntax: str) -> str:
     Returns:
         Path to the generated object file
     """
-    _, format_str = asm_common.get_asm_config(syntax)
+    _, format_str = get_asm_config(syntax)
 
     if output_path is None:
         name = asm_block.name or "asm_block"
@@ -300,8 +744,8 @@ def print_asm_block(asm_block, is_arm64: bool = False):
         for instr in instructions:
             if instr.strip().lower() == "ret":
                 continue
-            for expanded in asm_common._expand_write_call(instr, is_arm64=False):
-                for final_instr in asm_common._expand_len_instruction(expanded, is_arm64=False):
+            for expanded in _expand_write_call(instr, is_arm64=False):
+                for final_instr in _expand_len_instruction(expanded, is_arm64=False):
                     print(f"    {final_instr}")
         print("    ret")
     else:
@@ -316,6 +760,6 @@ def print_asm_block(asm_block, is_arm64: bool = False):
             if not "data" in directive.lower() and not directive.startswith("syntax"):
                 print(directive)
         for line in instructions:
-            for expanded in asm_common._expand_write_call(line, is_arm64=False):
-                for final_instr in asm_common._expand_len_instruction(expanded, is_arm64=False):
+            for expanded in _expand_write_call(line, is_arm64=False):
+                for final_instr in _expand_len_instruction(expanded, is_arm64=False):
                     print(final_instr)
